created endpoints and fixes errors

+++ src/main/java/core/ics/domain/usecase/impl/PersonMaintenanceUseCase.java
@@ -8,11 +8,11 @@ import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 @ApplicationScoped
-public class PersonMaintenanceUseCase implements IMaintenanceUseCase<Person,Long> {
+public class PersonMaintenanceUseCase implements IMaintenanceUseCase<Person,Integer> {
 
     @Inject
-    IGatewayBase<Person, Long> gateway;
+    IGatewayBase<Person, Integer> gateway;
 
     @Override
-    public IGatewayBase<Person, Long> getGateway() { return gateway; }
+    public IGatewayBase<Person, Integer> getGateway() { return gateway; }
 }
